[SIDE-461] Add initial TS config
