+++ src/connInfo.js
