+++ src/controllerGrpcServer.js
@@ -1,30 +1,31 @@
-const messages = require('./proto/ldk_pb');
-const services = require('./proto/ldk_grpc_pb');
+const messages = require('../proto/ldk_pb');
+const services = require('../proto/ldk_grpc_pb');
 
 const { categories } = require('./categories');
 
 const BrokerGrpcServer = require('./brokerGrpcServer');
-const SensorGrpcHostClient = require('./sensorGrpcHostClient');
-const Sensor = require('./sensor');
+const Controller = require('./controller');
+
+const ControllerGrpcHostClient = require('./controllerGrpcHostClient');
 
 /**
- * Class used by the host process to interact with the sensor implementation.
+ * Class used by the host process to interact with the controller implementation.
  *
  * @private
  */
-class SensorGRPCServer {
+class ControllerGrpcServer {
   /**
-   * Create a SensorGRPCServer.
+   * Create a ControllerGrpcServer.
    *
    * @param {object} server - The GRPC server instance.
-   * @param {Sensor} impl - The sensor implementation.
+   * @param {Controller} impl - The controller implementation.
    * @param {BrokerGrpcServer} broker - The GRPC broker server instance.
    * @example
-   * SensorGRPCServer(server, mySensor, broker);
+   * ControllerGrpcServer(server, myController, broker);
    */
   constructor(server, impl, broker) {
     this.broker = broker;
-    server.addService(services.SensorService, {
+    server.addService(services.ControllerService, {
       start: this.start(impl),
       stop: this.stop(impl),
       onEvent: this.onEvent(impl),
@@ -32,10 +33,10 @@ class SensorGRPCServer {
   }
 
   /**
-   * Called by the host to start the sensor implementation.
+   * Called by the host to start the controller implementation.
    *
    * @async
-   * @param {Sensor} impl - The implementation of the sensor.
+   * @param {Controller} impl - The implementation of the controller.
    * @returns {void}
    */
   start(impl) {
@@ -45,7 +46,7 @@ class SensorGRPCServer {
 
       const connInfo = await this.broker.getConnInfo();
 
-      const hostClient = new SensorGrpcHostClient();
+      const hostClient = new ControllerGrpcHostClient();
       await hostClient.connect(connInfo).catch((err) => {
         throw err;
       });
@@ -57,10 +58,10 @@ class SensorGRPCServer {
   }
 
   /**
-   * Called by the host to stop the sensor implementation.
+   * Called by the host to stop the controller implementation.
    *
    * @async
-   * @param {Sensor} impl - The implementation of the sensor.
+   * @param {Controller} impl - The implementation of the controller.
    * @returns {void}
    */
   stop(impl) {
@@ -73,10 +74,10 @@ class SensorGRPCServer {
   }
 
   /**
-   * Called by the host to broadcast events to the sensor implementation.
+   * Called by the host to broadcast events to the controller implementation.
    *
    * @async
-   * @param {Sensor} impl - The implementation of the sensor.
+   * @param {Controller} impl - The implementation of the controller.
    * @returns {void}
    */
   onEvent(impl) {
@@ -104,4 +105,4 @@ class SensorGRPCServer {
   }
 }
 
-module.exports = SensorGRPCServer;
+module.exports = ControllerGrpcServer;
