+++ src/sensor.js
