+++ src/sensorPlugin.js
@@ -1,6 +1,6 @@
 /** @module sensorPlugin */
 
-const services = require('./proto/ldk_grpc_pb');
+const services = require('../proto/ldk_grpc_pb');
 const { prepareLogging } = require('./logging');
 
 const BrokerGrpcServer = require('./brokerGrpcServer');
