+++ src/whisper.js
